Changed to generic launcher and indivudal apk builds
