+++ app/src/all4/res/mipmap-hdpi/icon.png
