+++ app/src/all4/res/mipmap-mdpi/icon.png
