+++ app/src/all4/res/mipmap-xhdpi/icon.png
