+++ app/src/all4/res/mipmap-xxhdpi/icon.png
