+++ app/src/iPlayer/res/mipmap-hdpi/icon.png
