+++ app/src/iPlayer/res/mipmap-mdpi/icon.png
