+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="hosh.io.atv_launchers">
+    xmlns:tools="http://schemas.android.com/tools"
+    package="hosh.io.atv_launchers">
 
     <uses-feature
             android:name="android.hardware.touchscreen"
@@ -11,47 +12,17 @@
 
     <application android:allowBackup="true"
                  android:supportsRtl="true"
-                 android:icon="@drawable/movie"
-                 android:theme="@style/Theme.Leanback">
-        <activity android:name="hosh.io.atv_launchers.BbcIplayerActivity"
-                  android:taskAffinity="hosh.io.atv_launchers.BbcIplayerActivity"
-                  android:excludeFromRecents="true"
-                  android:label="@string/bbc_iplayer_name"
-                  android:icon="@mipmap/bbc_iplayer_icon"
-                  android:banner="@mipmap/bbc_iplayer_banner">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN"/>
-                <category android:name="android.intent.category.LEANBACK_LAUNCHER"/>
-            </intent-filter>
-        </activity>
-        <activity android:name="hosh.io.atv_launchers.ItvHubActivity"
-                  android:taskAffinity="hosh.io.atv_launchers.ItvHubActivity"
-                  android:excludeFromRecents="true"
-                  android:label="@string/itv_hub_name"
-                  android:icon="@mipmap/itv_hub_icon"
-                  android:banner="@mipmap/itv_hub_banner">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN"/>
-                <category android:name="android.intent.category.LEANBACK_LAUNCHER"/>
-            </intent-filter>
-        </activity>
-        <activity android:name="hosh.io.atv_launchers.All4Activity"
-                  android:taskAffinity="hosh.io.atv_launchers.All4Activity"
-                  android:excludeFromRecents="true"
-                  android:label="@string/all_4_name"
-                  android:icon="@mipmap/all_4_icon"
-                  android:banner="@mipmap/all_4_banner">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN"/>
-                <category android:name="android.intent.category.LEANBACK_LAUNCHER"/>
-            </intent-filter>
-        </activity>
-        <activity android:name="hosh.io.atv_launchers.My5Activity"
-                  android:taskAffinity="hosh.io.atv_launchers.My5Activity"
-                  android:excludeFromRecents="true"
-                  android:label="@string/my5_name"
-                  android:icon="@mipmap/my5_icon"
-                  android:banner="@mipmap/my5_banner">
+                 android:icon="@mipmap/icon"
+                 android:label="@string/label"
+                 android:banner="@mipmap/banner"
+                 android:theme="@style/Theme.Leanback"
+                 tools:ignore="GoogleAppIndexingWarning">
+        <meta-data android:name="serviceUrl" android:value="@string/serviceUrl" />
+        <meta-data android:name="serviceName" android:value="@string/serviceName" />
+
+        <activity android:name="hosh.io.atv_launchers.GenericLauncher"
+                  android:taskAffinity="hosh.io.atv_launchers.GenericLauncher"
+                  android:excludeFromRecents="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LEANBACK_LAUNCHER"/>
